criacao de excecoes e tratamento de erros

+++ src/main/java/com/fullstack/education/labpcp/service/serviceImpl/CursoServiceImpl.java
@@ -6,6 +6,8 @@ import com.fullstack.education.labpcp.controller.dto.response.CursoResponse;
 import com.fullstack.education.labpcp.datasource.entity.CursoEntity;
 import com.fullstack.education.labpcp.datasource.entity.MateriaEntity;
 import com.fullstack.education.labpcp.datasource.repository.CursoRepository;
+import com.fullstack.education.labpcp.datasource.repository.TurmaRepository;
+import com.fullstack.education.labpcp.infra.exception.*;
 import com.fullstack.education.labpcp.service.CursoService;
 import com.fullstack.education.labpcp.service.TokenService;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,45 +24,71 @@ public class CursoServiceImpl implements CursoService {
 
     private final CursoRepository cursoRepository;
     private final TokenService tokenService;
+    private final TurmaRepository turmaRepository;
 
-    public void papelUsuarioAcessoPermitido(String token){
+    public void papelUsuarioAcessoPermitido(String token) {
 
         String nomePapel = tokenService.buscaCampo(token, "scope");
 
-        if(Objects.equals(nomePapel, "RECRUITER") || Objects.equals(nomePapel, "PROFESSOR") || Objects.equals(nomePapel, "ALUNO")){
-            throw new RuntimeException("Apenas administradores e pedagogos podem acessar os endpoints de curso!");
+        if (Objects.equals(nomePapel, "RECRUITER") || Objects.equals(nomePapel, "PROFESSOR") || Objects.equals(nomePapel, "ALUNO")) {
+            throw new AcessoNaoAutorizadoException("Apenas administradores e pedagogos podem acessar os endpoints de curso!");
         }
     }
 
     @Override
     public CursoResponse criarCurso(CursoRequest cursoRequest, String token) {
 
+        if (cursoRequest.nome() == null) {
+            throw new CampoAusenteException("O campo nome é obrigatório!");
+        }
+
+
+        Optional<CursoEntity> cursoExistenteOptional = cursoRepository.findByNome(cursoRequest.nome().toUpperCase());
+        if (cursoExistenteOptional.isPresent()) {
+            throw new RegistroExistenteException("Já existe curso cadastrado com este nome!");
+        }
+
+
         papelUsuarioAcessoPermitido(token);
         CursoEntity curso = new CursoEntity();
-        curso.setNome(cursoRequest.nome());
+        curso.setNome(cursoRequest.nome().toUpperCase());
         cursoRepository.save(curso);
         return new CursoResponse(curso.getId(), curso.getNome());
     }
 
-    public CursoEntity cursoPorId(Long id){
-        CursoEntity cursoPesquisado = cursoRepository.findById(id).orElseThrow(() -> new RuntimeException("Id não correspondente a nenhum curso cadastrado"));
+    public CursoEntity cursoPorId(Long id) {
+        CursoEntity cursoPesquisado = cursoRepository.findById(id).orElseThrow(() -> new NotFoundException("Id não correspondente a nenhum curso cadastrado"));
         return cursoPesquisado;
     }
+
     @Override
     public CursoResponse obterCursoPorId(Long id, String token) {
         papelUsuarioAcessoPermitido(token);
         CursoEntity cursoPesquisado = cursoPorId(id);
-        return new CursoResponse(cursoPesquisado.getId(),cursoPesquisado.getNome());
+        return new CursoResponse(cursoPesquisado.getId(), cursoPesquisado.getNome());
     }
 
     @Override
     public CursoResponse atualizarCurso(Long id, CursoRequest cursoRequest, String token) {
         papelUsuarioAcessoPermitido(token);
+
+        if (cursoRequest.nome() == null) {
+            throw new CampoAusenteException("O campo nome é obrigatório!");
+
+        }
         CursoEntity cursoPesquisado = cursoPorId(id);
+
+        if(!cursoPesquisado.getNome().equalsIgnoreCase(cursoRequest.nome().toUpperCase())){
+            Optional<CursoEntity> cursoExistenteOptional = cursoRepository.findByNome(cursoRequest.nome().toUpperCase());
+            if (cursoExistenteOptional.isPresent()) {
+                throw new RegistroExistenteException("Já existe curso cadastrado com este nome!");
+            }
+        }
+
         cursoPesquisado.setId(id);
-        cursoPesquisado.setNome(cursoRequest.nome());
+        cursoPesquisado.setNome(cursoRequest.nome().toUpperCase());
         cursoRepository.save(cursoPesquisado);
-        return new CursoResponse(cursoPesquisado.getId(),cursoPesquisado.getNome());
+        return new CursoResponse(cursoPesquisado.getId(), cursoPesquisado.getNome());
     }
 
     @Override
@@ -67,11 +96,16 @@ public class CursoServiceImpl implements CursoService {
 
         String nomePapel = tokenService.buscaCampo(token, "scope");
 
-        if(!Objects.equals(nomePapel, "ADM")){
-            throw new RuntimeException("Apenas administradores podem podem excluir curso!");
+        if (!Objects.equals(nomePapel, "ADM")) {
+            throw new AcessoNaoAutorizadoException("Apenas administradores podem podem excluir curso!");
         }
 
         CursoEntity cursoPesquisado = cursoPorId(id);
+        String nomeCursoPesquisado = cursoPesquisado.getNome();
+        boolean cursoAssociadoAUmaTurma = turmaRepository.existsByNomeCursoNome(nomeCursoPesquisado);
+        if(cursoAssociadoAUmaTurma){
+            throw new BadRequestException("Este curso está associado a pelo menos uma turma, indique outro curso a turma antes de deleta-lo!");
+        }
         cursoRepository.delete(cursoPesquisado);
 
     }
